Resource SQL Mixins
This patch adds a couple of improvements for universal agent topic. The
main improvements is `TargetResourceSQLStorableMixin` with universal
methods to get new, updated and deleted entities.
The second part is ability to work with DB (PG) for the agent. So any
drivers that need DB access may include in the configuration a common
'db' section.
Also a coule of bugs were fixed.

Signed-off-by: Anton Kremenetsky <anton.kremenetsky@gmail.com>