Unique res_uuid based on uuid5(uuid, kind) and Status API adaptation
The target and actual UA resources unable to have only UUID as a primary
key since we can have the same UUID but different kind. For instance, we
can reach this if create a config via Element Manager. After creating
two resources with the same UUID but different kind (config,
em_core_config) are created. So an unique uuid `res_uuid` is introduced
to solve this problem. The `res_uuid` may be any UUID but it's highly
recommended to use `uuid5(uuid, kind)`, the helpful methods were added.
In this way there are two UUIDs:
- `uuid`. Should be the same as `uuid` of internal object that the resource
is holding.
- `res_uuid`. It's an unique UUID of resource, it's highly recommended
to use `uuid5(uuid, kind)` for it.

`res_uuid` is mostly used only in database and RA context to correct
making primary, foreign keys and other relations. `uuid` is used around
the code base, so it's recommended to used it in the development.

Also it requires additional adaptation of Status API since we cannot
fetch a resource by the `/v1/resource/<uuid>` endpoint because of the
same reason. Therefore resources have been moved under
`/v1/kind/<kind-name>/resources/<uuid>` endpoint.

The `kind` collection is dynamic:
```bash
curl http://10.20.0.2:11012/v1/kind/ | jq .
[]

curl http://10.20.0.2:11012/v1/kind/foobar/resources/ | jq .
[]
```

If any resource is added in a new `kind`, this kind will be shown in the
`/v1/kind/` after this moment.

Other improvements:
- Added tests for Status API

Signed-off-by: Anton Kremenetsky <anton.kremenetsky@gmail.com>