Integration with TS storage for backend client
The backend clients `DatabaseBackendClient` and `GCRestApiBackendClient`
were integrated with `target fields storage` to have an ability for
flexble filtering when there is no a specific field like `project_id`.
In this case they use collection of UUIDs from the `target fields storage`
to detect which entities belongs them and which not.

Signed-off-by: Anton Kremenetsky <anton.kremenetsky@gmail.com>